+++ NEWS.md
@@ -4,7 +4,7 @@ will continue until then.  These will be especially noticable around
 the inclusion of IV NCA parameters and additional specifications of
 the dosing including dose amount and route.
 
-# PKNCA 0.9.0
+# PKNCA 0.8.5
 
 * Cleaned AUCint names
 * Added dose-count within interval (to warn of multiple doses within an